* bug.kaoni.com(#17687) 동영상 게시판 1.6GB 파일 첨부되지 않음
동영상 게시판 수정(personalizedPortal)

+++ src/main/java/egovframework/ezEKP/ezBoard/web/EzBoardController.java
@@ -8918,7 +8918,7 @@ public class EzBoardController extends EgovFileMngUtil{
 			pFileLimit = "2";
 		}
 		
-		int fileLimit = Integer.parseInt(pFileLimit) * 1024 * 1024;
+		long fileLimit = Long.parseLong(pFileLimit) * 1024 * 1024;
 		
 		fileSize = multiFile.getSize();
 		if (fileSize > fileLimit) {
